🎛️🎨: extract `VideoMorph` into `lively.components`
