+++ lively.morphic/html-morph.js
@@ -271,237 +271,3 @@ export class IFrameMorph extends HTMLMorph {
     return '<iframe width="100%" height="100%" frameBorder="false" srcdoc=""></iframe>';
   }
 }
-
-/**
- * A VideoMorph, which is just an overlay on the HTML Video element
- * This is a subclass of the HTMLMorph, and so it inherits all of the HTMLMorph's
- * properties.
- * properties (should be in the constructor)
- *   - src: url for the video
- *   - type: type of the video (mp4, ogg, etc)
- *   - controls: true/false to show not/show the controls (default false)
- *   - loop: true/false to play on a loop (default true)
- *   - autoplay: true/false to play on load (default true)
- *   - videoLayout:  one of 'autosize', 'autoaspect', 'none'. Default to autoaspect.
- *          If 'autosize', size the morph to the video's size.
- *          If 'autoaspect', resize to correct aspect ratio within current bounding box
- * methods:
- *    - startPlaying() -- start playing the video
- *    - stopPlaying() -- stop playing the video
- *    - rewind()  -- set the video counter to 0
- *
- * read-only properties:
- *     - videoDomElement -- the DOM element for the video
- */
-
-// new VideoMorph({autoplay: true, src:'https://matt.engagelively.com/assets/ITF/New-creative-video-6-.mp4', videoLayout:'autosize'}).openInWorld()
-export class VideoMorph extends HTMLMorph {
-  static get properties () {
-    return {
-      autoplay: {
-        defaultValue: false,
-        set (aBool) {
-          this.setVideoProperty('autoplay', !!aBool);
-        }
-      },
-      loop: {
-        defaultValue: false,
-        set (aBool) {
-          this.setVideoProperty('loop', !!aBool);
-        }
-      },
-      controls: {
-        defaultValue: false,
-        set (aBool) {
-          this.setVideoProperty('controls', !!aBool);
-        }
-      },
-      src: {
-        defaultValue: 'https://matt.engagelively.com/assets/kaleidoscope-art-17141.mp4',
-        set (srcURL) {
-          this.setVideoProperty('src', srcURL);
-          this.videoLoaded = false;
-        }
-      },
-      type: {
-        defaultValue: 'video/mp4',
-        set (videoType) {
-          this.setVideoProperty('type', videoType);
-        }
-      },
-      videoLayout: {
-        defaultValue: 'autoaspect',
-        set (aLayoutStyle) {
-          // console.log('Setting videoLayout to ' + aLayoutStyle);
-          const choices = ['none', 'autoaspect', 'autosize'];
-          const index = choices.indexOf(aLayoutStyle);
-          const choice = index >= 0 ? choices[index] : 'none';
-          this.setProperty('videoLayout', choice);
-          this.resizeToExtentProperties();
-        }
-
-      }
-
-    };
-  }
-
-  constructor (props) {
-    // this.videoId = `video-${Math.floor(Math.random() * 1E6)}`;
-    this.constructorProps = props;
-    super(props);
-  }
-
-  setVideoProperty (propertyName, value) {
-    this.setProperty(propertyName, value);
-    this.resetHTML();
-  }
-
-  onLoad () {
-    const props = this.constructorProps;
-    this.log = [];
-    this.autorelayout = true;
-    this.videoLoaded = false;
-    const setField = (field, defaultVal) => {
-      this[field] = props && props[field] ? props[field] : defaultVal;
-      this.log.push([field, this[field]]);
-    };
-    const normalizeBoolean = field => {
-      this[field] = props ? !!props[field] : false;
-    };
-
-    const bools = ['loop', 'controls', 'autoplay'];
-    bools.forEach(bool => normalizeBoolean(bool));
-    const fields = [
-      { name: 'src', defaultVal: 'https://matt.engagelively.com/assets/kaleidoscope-art-17141.mp4' },
-      { name: 'type', defaultVal: 'video/mp4' },
-      { name: 'badBrowserMessage', defaultVal: 'Your browser does not support video content' },
-      { name: 'loop', defaultVal: false },
-      { name: 'autoplay', defaultVal: false },
-      { name: 'controls', defaultVal: false }
-      // { name: 'videoLayout', defaultVal: 'autoaspect' }
-    ];
-
-    fields.forEach(field => setField(field.name, field.defaultVal));
-    this.resetHTML();
-
-    this.awaitingResize = false;
-
-    this.autorelayout = false;
-  }
-
-  resizeToExtentProperties () {
-    // resize to the extent properties.  Since the default behavior on
-    // bounds change is to set videoLayout to 'none', and we don't want that
-    // to happen in this case, set this.autorelayout to true to tell
-    // onBoundsChanged() not to reset the videoLayout method
-    this.whenRendered().then(_ => {
-      // console.log('in resizeToExtentProperties, this.videoLayout = ' + this.videoLayout);
-
-      this.autorelayout = true;
-      if (this.videoLayout === 'autosize') {
-        // console.log('Resizing to natural extent');
-        this.resizeToNaturalExtent();
-      } else if (this.videoLayout === 'autoaspect') {
-        // console.log('Resizing to fit video');
-        this.resizeToFitVideo();
-      }
-
-      this.autorelayout = false;
-      // console.log('leaving resizeToExtentProperties');
-    });
-  }
-
-  get domElementExtent () {
-    return pt(this.videoDomElement.width, this.videoDomElement.height);
-  }
-
-  onBoundsChanged (bounds) {
-    super.onBoundsChanged(bounds);
-    // console.log('In onBoundsChanged');
-    if (!this.autorelayout) {
-      // once a user has resized, this turns off autosizing to accomodate the video
-      // the autorelayout flag is set by the resize methods to ensure that this
-      // won't happen when we are resizing to accomodate a video
-      this.videoLayout = 'none';
-    }
-
-    if (this.bounds().extent().eqPt(this.domElementExtent)) {
-      return;
-    }
-    // console.log('Leaving onBoundsChanged');
-    this.resetHTML();
-  }
-
-  init () {
-    this.resetHTML();
-    this.startPlaying();
-  }
-
-  get videoDomElement () {
-    return this.domNode.querySelectorAll('video')[0];
-  }
-
-  get naturalExtent () {
-    if (this.videoDomElement) {
-      return pt(this.videoDomElement.videoWidth, this.videoDomElement.videoHeight);
-    } else {
-      return pt(10, 10);
-    }
-  }
-
-  resizeToNaturalExtent () {
-    if (this.videoLoaded) {
-      // console.log('natural extent is' + this.naturalExtent);
-      this.extent = this.naturalExtent;
-    } else {
-      // console.log('Video is not yet loaded');
-    }
-  }
-
-  resizeToFitVideo () {
-    const videoExtent = this.naturalExtent;
-    const widthRatio = this.width / videoExtent.x;
-    const heightRatio = this.height / videoExtent.y;
-    // remain within the current bounding box, but shrink one dimension so
-    // this.width/this.height = video.extent.x/video.extent.y
-    if (heightRatio > widthRatio) {
-      // morph is too tall for the video
-      this.extent = pt(this.extent.x, videoExtent.y * widthRatio);
-    } else if (widthRatio > heightRatio) {
-      // morph is too wide for the video
-      this.extent = pt(videoExtent.x * heightRatio, this.extent.y);
-    }
-  }
-
-  rewind () {
-    this.videoDomElement.currentTime = 0;
-  }
-
-  startPlaying () {
-    this.videoDomElement.play();
-  }
-
-  stopPlaying () {
-    this.videoDomElement.pause();
-    this.videoDomElement.currentTime = 0;
-  }
-
-  resetHTML () {
-    const options = ` ${this.loop ? 'loop' : ''} ${this.controls ? 'controls' : ''} ${this.autoplay ? 'autoplay' : ''}`;
-    this.html = `
-  <video id="${this.videoId}"  this.width = "${this.width}" height="${this.height}" ${options}>
-  <source src="${this.src}"  type="${this.type}"/> 
-${this.badBrowserMessage}
-</video>
-`;
-  }
-
-  menuItems () {
-    let items = super.menuItems();
-    // pop the first two items off the list.  Note this must change if
-    // HTMLMorph.menuItems() changes
-    // items.shift(); // remove edit html
-    // items.shift(); // remove edit css
-    return items;
-  }
-}
